feat: add new index page
